models/pyproject.toml
[project]
name = "analyze"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "matv864", email = "matv864@gmail.com" }
]
dependencies = [
    "prophet>=1.1.6",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.38",
    "psycopg2>=2.9.10",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/analyze"]
